Devices refactored to servers
