+++ lib/fog/compute/packet/models/devices.rb
@@ -4,18 +4,8 @@ module Fog
   module Compute
     class Packet
       # Devices Collection
-      class Devices < Fog::Collection
+      class Devices < Fog::Compute::Packet::Servers
         model Fog::Compute::Packet::Device
-
-        def all(project_id, params = {})
-          response = service.list_devices(project_id, params)
-          load(response.body["devices"])
-        end
-
-        def get(id)
-          response = service.get_device(id)
-          new(response.body)
-        end
       end
     end
   end
