+++ tests/packet/models/test_virtual_networks.rb
@@ -12,10 +12,10 @@ class TestVirtualNetworks < Minitest::Test
   end
 
   def test_a_create_virtual_network
-    device = @compute.devices.create(:project_id => @project_id, :facility => "ewr1", :plan => "baremetal_0", :hostname => "test01", :operating_system => "coreos_stable")
-    @@device_id = device.id
+    server = @compute.servers.create(:project_id => @project_id, :facility => "ewr1", :plan => "baremetal_0", :hostname => "test01", :operating_system => "coreos_stable")
+    @@server_id = server.id
 
-    device.wait_for { ready? } if Fog.mock?
+    server.wait_for { ready? } if Fog.mock?
 
     response = @compute.virtual_networks.create(:project_id => @project_id, :description => "test", :facility => "ewr1", :vlan => 1, :vxlan => 1)
 
@@ -42,8 +42,8 @@ class TestVirtualNetworks < Minitest::Test
 
   def test_f_assign_port
     eth1 = ""
-    device = @compute.devices.get(@@device_id)
-    device.provisioning_events.each do |port|
+    server = @compute.servers.get(@@server_id)
+    server.provisioning_events.each do |port|
       next unless port["network_ports"]
       port["network_ports"].each do |np|
         eth1 = np["id"] if np["name"] == "eth1"
@@ -56,8 +56,8 @@ class TestVirtualNetworks < Minitest::Test
 
   def test_g_unassign_port
     eth1 = ""
-    device = @compute.devices.get(@@device_id)
-    device.provisioning_events.each do |port|
+    server = @compute.servers.get(@@server_id)
+    server.provisioning_events.each do |port|
       next unless port["network_ports"]
       port["network_ports"].each do |np|
         eth1 = np["id"] if np["name"] == "eth1"
@@ -75,7 +75,7 @@ class TestVirtualNetworks < Minitest::Test
   end
 
   def test_i_cleanup
-    device = @compute.devices.get(@@device_id)
-    device.destroy
+    server = @compute.servers.get(@@server_id)
+    server.destroy
   end
 end
